fix routing for charts
